factored out list hide functionality from render portion of QuestionList component, to its own bound component function named hide
